Include a new node and update infra
This commit introduces a new machine on the setup. Doing that on a new computer required new versions of the Terraform plugins

This commit removes the forked versions of Cloudflare and ACME by using the latest version with the patches applied.

It required a new forked patch on Zerotier provider to run it with Terraform 0.12.24

All changes applied, including new certificates